Fix bug where multi-line strings were causing errors with print().
The REGEX wasn't removing the space before the gutter for
mutliline strings, so I fixed that. The bigger issue turned out
to be that I was using the multiline flag for a REGEX that was
checking for a single-line variable, and this caused it to
fail when it encountered a mutliline string. I had to remove a
flag from Posiedon that was there too, as it was always null
so wasn't doing anything.